API

API

ArcSegment

API

Clothoid

SmoothCurves.jl

Curve Types

Curve
ArcSegment

SmoothCurves.Curve — Type.

Abstract base class for all specialized curve types.

source

SmoothCurves.ArcSegment — Type.

ArcSegment(origin, radius, angle0, angle1)

ArcSegment centered at origin with the given radius going from the first to the second angle. As parametrization the angle difference to the starting angle is used.

source

Curve Interface

smax
point
dpoint
length
dlength
curvature
dcurvature
tangentangle
radialangle

SmoothCurves.smax — Function.

smax(C::Curve)

Maximum parameter $s$ for which the given curve $\mathcal{C}$ is defined.

source

SmoothCurves.point — Function.

point(C::Curve, s)

Point (x, y) of the curve $\mathcal{C}$ at $s$. Returns a StaticArrays.SVector.

source

SmoothCurves.dpoint — Function.

dpoint(C::Curve, s)

Derivative $\frac{dC(s)}{ds}$ of the curve $\mathcal{C}$ at $s$. Returns a StaticArrays.SVector.

source

Base.length — Function.

length(C::Curve, s)

Length of the given curve $\mathcal{C}$ between $0$ and $s$.

source

length(C::Curve)

Total length of the given curve. Equal to length(C, smax(C)).

source

SmoothCurves.dlength — Function.

dlength(C::Curve, s)

Derivative $\frac{dl}{ds}$ at $s$.

source

SmoothCurves.curvature — Function.

curvature(C::Curve, s)

Curvature of the Curve $\mathcal{C}$ at $s$.

source

SmoothCurves.dcurvature — Function.

dcurvature(C::Curve, s)

Derivative $\frac{dκ}{ds}$ of the curve $\mathcal{C}$ at $s$.

source

SmoothCurves.tangentangle — Function.

tangentangle(C::Curve, s)

Angle between the tangent of the curve $\mathcal{C}$ at $s$ and the $x$-axis. Always between $-π$ and $π$.

source

SmoothCurves.radialangle — Function.

radialangle(C::Curve, s)

Angle between the radial vector of the curve $\mathcal{C}$ at $s$ and the $x$-axis. Always between $-π$ and $π$.

source

Curve specific functions

ArcSegment

Base.sign — Method.

sign(C::ArcSegment)

Sign of the curvature of the ArcSegment.

source